Assignment 3
Reflection on assignment 3 :
The UTMDigital visit proved to be a captivating experience for us, particularly as students of computing. It provided us with a profound insight into the distinctive advancements in technology and information systems that impact our daily lives. Throughout the visit, we had the opportunity to familiarize ourselves with the roles of UTMDigital staff, who play a crucial role in maintaining UTM's digital ICT services. Notable innovations like UTMWifi and the MyUTM Portal were highlighted, showcasing the transformative effects of technology on online services.