Assignment 2
Reflection on Assignment 2 : 
The experience with Clarity Techwork was truly remarkable. Mr. Yee Soon Tuck, the Chief Technology Officer (CTO) of the company, delivered an inspiring talk that left a lasting impression. One noteworthy aspect was his revelation that he commenced his IT career without holding a degree. This serves as a reminder that securing employment requires not only theoretical knowledge but also practical skills. Moreover, the significance of attitude was emphasized, highlighting its crucial role in thriving within a work environment. We perceive this talk as a valuable opportunity for students to gain insights into their future careers, enabling them to enhance their collaborative skills and sense of responsibility.